+++ xla/client_test.go
@@ -34,18 +34,5 @@ func TestBuildClient(t *testing.T) {
 			t.Errorf("\tFailed to create Client for %q: %v", platform, err)
 			continue
 		}
-		/*
-			if !o.Run() {
-				t.Errorf("\tOrchestrator %q failed.", name)
-				continue
-			}
-			shape, data := o.Result()
-			wantShape := []int64{3}
-			wantData := []float32{10.5, 21, 32}
-			fmt.Printf("\tGot: Shape=%v, Flat=%v\n", shape, data)
-			if !reflect.DeepEqual(shape, wantShape) || !reflect.DeepEqual(data, wantData) {
-				fmt.Printf("Want: Shape=%v, Flat=%v\n", wantShape, wantData)
-			}
-		*/
 	}
 }